requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
httpx>=0.27,<1
numpy>=1.26,<2
Pillow
pycocotools
einops
decord
opencv-python-headless
scipy
scikit-image
rapidfuzz
paddlepaddle==3.2.2
paddleocr==3.4.0
easyocr
